test for fine_amplitude with fixture
